Move is_in_countrylist() to score.c
